# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fund Daily API",
    "version": "2.5.1",
    "description": "基金每日分析系统 API 文档",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "开发服务器"
    }
  ],
  "paths": {
...
    "/api/funds": {
      "get": {
        "summary": "",
        "description": "获取用户持有的基金列表或默认基金列表",
        "tags": [
          "基金数据"
        ],
        "responses": {
          "200": {
            "description": "基金列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "funds": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Fund"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "返回数量限制",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          }
        ]
      }
    },
    "/api/fund-detail/{code}": {
      "get": {
        "summary": "",
        "description": "根据基金代码获取基金详细信息",
        "tags": [
          "基金数据"
        ],
        "responses": {
          "200": {
            "description": "基金详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "detail": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "description": "基金代码（6位数字）",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{6}$"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Fund": {
        "type": "object",
        "properties": {
          "fund_code": {
            "type": "string",
            "example": "000001"
          },
          "fund_name": {
            "type": "string",
            "example": "华夏成长"
          },
          "net_value": {
            "type": "number",
            "example": 1.2345
          },
          "daily_change": {
            "type": "number",
            "example": 1.23
          },
          "return_1y": {
            "type": "number",
            "example": 15.5
          },
          "risk_level": {
            "type": "string",
            "example": "中高风险"
          },
          "trend": {
            "type": "string",
            "example": "up"
          },
          "score": {
            "type": "number",
            "example": 85.5
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  },
...
}